requirements.txt
numpy==1.20.1
streamlit==0.76.0
Pillow==8.1.0